% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pinpoint_data.R
\name{readPinpoint}
\alias{readPinpoint}
\title{readPinpoint}
\usage{
readPinpoint(file, birdID = NULL, band = NULL, database = NULL,
  start, stop, breedyear = NULL)
}
\arguments{
\item{file}{String, filename to read in.}

\item{birdID}{String, optional ID of bird to associate with records}

\item{band}{String, optional band number of bird to associate with records}

\item{database}{Optional, object name of existing database to append with
read in data.}

\item{start}{String, starting date/time of tracking data to retain.  Format:
"y/m/d h:m:s"}

\item{stop}{String, ending date/time of tracking data to retain.  Format:
"y/m/d h:m:s"}

\item{breedyear}{String, optional year of data to associate with records}
}
\description{
readPinpoint
}
